+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Upgrade to NYPL Design System version 1.0.0
 - Decouples search query into separate states for each individual form field
 - Updates unit tests and snapshots
+- CHORE: removes tugboat deploy
+- Change format filters to "Readable", "Downloadable", and "Requestable"
 
 ## [0.13.1]
 - Added "Back to search results" link for works and editions